🎨  Use layout="fixed" for small media types
no issue

Using `layout="responsive"` as a default value might end in ugly results. Small images get stretched to fit in the parent container. Viewport for AMP should vary between 600px and 300px. In case, we have a media type smaller then 300px, we don't wont to stretch it, so we set the fixed layout and AMP will use the detected media dimensions.